Extract useRoutes into common Squid lib
