+++ packages/checkout/widgets-lib/src/lib/squid/getDurationFormatted.ts
