+++ packages/checkout/widgets-lib/src/lib/squid/getTotalRouteFees.ts
