+++ packages/checkout/widgets-lib/src/lib/squid/isSquidNativeToken.ts
@@ -1,3 +1,3 @@
-import { SQUID_NATIVE_TOKEN } from '../utils/config';
+import { SQUID_NATIVE_TOKEN } from './config';
 
 export const isSquidNativeToken = (token: string) => token.toLowerCase() === SQUID_NATIVE_TOKEN.toLowerCase();
