+++ packages/checkout/widgets-lib/src/widgets/add-tokens/components/Options.tsx
@@ -6,7 +6,8 @@ import { Checkout } from '@imtbl/checkout-sdk';
 import { useTranslation } from 'react-i18next';
 import { listItemVariants, listVariants } from '../../../lib/animation/listAnimation';
 import { FiatOption } from './FiatOption';
-import { Chain, FiatOptionType, RouteData } from '../types';
+import { FiatOptionType } from '../types';
+import { Chain, RouteData } from '../../../lib/squid/types';
 import { RouteOption } from './RouteOption';
 import { getRemoteVideo } from '../../../lib/utils';
