+++ packages/checkout/widgets-lib/src/widgets/add-tokens/utils/config.ts
@@ -1,7 +1,3 @@
-export const SQUID_SDK_BASE_URL = 'https://apiplus.squidrouter.com';
-
-export const SQUID_NATIVE_TOKEN = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE';
-
 export const FIXED_HANDOVER_DURATION = 2000;
 
 export const APPROVE_TXN_ANIMATION = '/access_coins.riv';
